requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
openai
gradio==4.44.1